# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/personalities": {
      "get": {
        "operationId": "PersonalitiesController_getPersonalities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List all personalities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonalityDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "personalities"
        ]
      },
      "post": {
        "operationId": "PersonalitiesController_createPersonality",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created personality",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalityDto"
                }
              }
            }
          }
        },
        "tags": [
          "personalities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/personalities/{id}": {
      "put": {
        "operationId": "PersonalitiesController_updatePersonality",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated personality",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonalityDto"
                }
              }
            }
          }
        },
        "tags": [
          "personalities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PersonalitiesController_deletePersonality",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personality deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "personalities"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/models/custom-selection/{userId}": {
      "get": {
        "operationId": "ModelsController_getCustomSelection",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get custom model selection for user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomSelectionWithModelsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "models"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "ModelsController_updateCustomSelection",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomModelSelectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update custom model selection for user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomSelectionWithModelsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "models"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ModelsController_deleteCustomSelection",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete custom model selection for user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsMessageResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "models"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Classroom Conversation Partner API",
    "description": "NestJS powered API with OpenAPI documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PersonalityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Personality ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "name": {
            "type": "string",
            "description": "Personality name"
          },
          "age": {
            "type": "number",
            "description": "Age",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar URL",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "description": "Gender"
          },
          "sex": {
            "type": "string",
            "description": "Sex",
            "enum": [
              "F",
              "M"
            ]
          },
          "voiceInstructions": {
            "type": "string",
            "description": "Voice instructions",
            "nullable": true
          },
          "elevenlabsVoiceId": {
            "type": "string",
            "description": "ElevenLabs voice ID",
            "nullable": true
          },
          "openaiVoiceName": {
            "type": "string",
            "description": "OpenAI voice name",
            "enum": [
              "alloy",
              "ash",
              "ballad",
              "coral",
              "echo",
              "sage",
              "shimmer"
            ]
          },
          "problemSummaryEn": {
            "type": "string",
            "description": "Problem summary in English"
          },
          "personalityDescriptionEn": {
            "type": "string",
            "description": "Personality description in English"
          },
          "problemSummaryCs": {
            "type": "string",
            "description": "Problem summary in Czech"
          },
          "personalityDescriptionCs": {
            "type": "string",
            "description": "Personality description in Czech"
          },
          "isHidden": {
            "type": "boolean",
            "description": "Whether personality is hidden"
          }
        },
        "required": [
          "id",
          "createdAt",
          "name",
          "age",
          "avatarUrl",
          "gender",
          "sex",
          "voiceInstructions",
          "elevenlabsVoiceId",
          "openaiVoiceName",
          "problemSummaryEn",
          "personalityDescriptionEn",
          "problemSummaryCs",
          "personalityDescriptionCs",
          "isHidden"
        ]
      },
      "CreatePersonalityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Personality name"
          },
          "age": {
            "type": "number",
            "description": "Age of the personality"
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar URL"
          },
          "gender": {
            "type": "string",
            "description": "Gender"
          },
          "sex": {
            "type": "string",
            "description": "Sex"
          },
          "voiceInstructions": {
            "type": "string",
            "description": "Voice instructions"
          },
          "elevenlabsVoiceId": {
            "type": "string",
            "description": "ElevenLabs voice ID"
          },
          "openaiVoiceName": {
            "type": "string",
            "description": "OpenAI voice name"
          },
          "problemSummaryEn": {
            "type": "string",
            "description": "Problem summary in English"
          },
          "personalityDescriptionEn": {
            "type": "string",
            "description": "Personality description in English"
          },
          "problemSummaryCs": {
            "type": "string",
            "description": "Problem summary in Czech"
          },
          "personalityDescriptionCs": {
            "type": "string",
            "description": "Personality description in Czech"
          },
          "isHidden": {
            "type": "boolean",
            "description": "Whether personality is hidden"
          }
        }
      },
      "UpdatePersonalityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Personality name"
          },
          "age": {
            "type": "number",
            "description": "Age of the personality"
          },
          "avatarUrl": {
            "type": "string",
            "description": "Avatar URL"
          },
          "gender": {
            "type": "string",
            "description": "Gender"
          },
          "sex": {
            "type": "string",
            "description": "Sex"
          },
          "voiceInstructions": {
            "type": "string",
            "description": "Voice instructions"
          },
          "elevenlabsVoiceId": {
            "type": "string",
            "description": "ElevenLabs voice ID"
          },
          "openaiVoiceName": {
            "type": "string",
            "description": "OpenAI voice name"
          },
          "problemSummaryEn": {
            "type": "string",
            "description": "Problem summary in English"
          },
          "personalityDescriptionEn": {
            "type": "string",
            "description": "Personality description in English"
          },
          "problemSummaryCs": {
            "type": "string",
            "description": "Problem summary in Czech"
          },
          "personalityDescriptionCs": {
            "type": "string",
            "description": "Personality description in Czech"
          },
          "isHidden": {
            "type": "boolean",
            "description": "Whether personality is hidden"
          }
        }
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ResponseModelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Model ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "friendlyName": {
            "type": "string",
            "description": "Friendly name"
          },
          "apiName": {
            "type": "string",
            "description": "API name"
          },
          "docsUrl": {
            "type": "string",
            "description": "Documentation URL",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether model is enabled"
          },
          "provider": {
            "type": "string",
            "description": "Model provider",
            "enum": [
              "OpenAi",
              "xAi",
              "Anthropic"
            ]
          }
        },
        "required": [
          "id",
          "createdAt",
          "friendlyName",
          "apiName",
          "isEnabled",
          "provider"
        ]
      },
      "TtsModelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Model ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "friendlyName": {
            "type": "string",
            "description": "Friendly name"
          },
          "apiName": {
            "type": "string",
            "description": "API name"
          },
          "sampleRate": {
            "type": "number",
            "description": "Sample rate"
          },
          "docsUrl": {
            "type": "string",
            "description": "Documentation URL"
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether model is enabled"
          },
          "provider": {
            "type": "string",
            "description": "Model provider",
            "enum": [
              "OpenAi",
              "ElevenLabs"
            ]
          },
          "allowsWordLevelTimestampedTranscript": {
            "type": "boolean",
            "description": "Whether model allows word-level timestamped transcripts"
          }
        },
        "required": [
          "id",
          "createdAt",
          "friendlyName",
          "apiName",
          "sampleRate",
          "docsUrl",
          "isEnabled",
          "provider",
          "allowsWordLevelTimestampedTranscript"
        ]
      },
      "RealtimeModelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Model ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "friendlyName": {
            "type": "string",
            "description": "Friendly name"
          },
          "apiName": {
            "type": "string",
            "description": "API name"
          },
          "docsUrl": {
            "type": "string",
            "description": "Documentation URL",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether model is enabled"
          },
          "provider": {
            "type": "string",
            "description": "Model provider",
            "enum": [
              "OpenAi"
            ]
          }
        },
        "required": [
          "id",
          "createdAt",
          "friendlyName",
          "apiName",
          "isEnabled",
          "provider"
        ]
      },
      "RealtimeTranscriptionModelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Model ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "friendlyName": {
            "type": "string",
            "description": "Friendly name"
          },
          "provider": {
            "type": "string",
            "description": "Model provider",
            "enum": [
              "OpenAi"
            ]
          },
          "apiName": {
            "type": "string",
            "description": "API name"
          },
          "docsUrl": {
            "type": "string",
            "description": "Documentation URL",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether model is enabled",
            "nullable": true
          },
          "allowsWordLevelTimestamps": {
            "type": "boolean",
            "description": "Whether model allows word-level timestamps"
          }
        },
        "required": [
          "id",
          "createdAt",
          "friendlyName",
          "provider",
          "apiName",
          "allowsWordLevelTimestamps"
        ]
      },
...
      "CustomSelectionWithModelsResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at timestamp"
          },
          "responseModelId": {
            "type": "number",
            "description": "Response model ID",
            "nullable": true
          },
          "ttsModelId": {
            "type": "number",
            "description": "TTS model ID",
            "nullable": true
          },
          "realtimeModelId": {
            "type": "number",
            "description": "Realtime model ID",
            "nullable": true
          },
          "realtimeTranscriptionModelId": {
            "type": "number",
            "description": "Realtime transcription model ID",
            "nullable": true
          },
          "timestampedTranscriptionModelId": {
            "type": "number",
            "description": "Timestamped transcription model ID",
            "nullable": true
          },
          "responseModel": {
            "description": "Response model",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ResponseModelResponseDto"
              }
            ]
          },
          "ttsModel": {
            "description": "TTS model",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TtsModelResponseDto"
              }
            ]
          },
          "realtimeModel": {
            "description": "Realtime model",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RealtimeModelResponseDto"
              }
            ]
          },
          "realtimeTranscriptionModel": {
            "description": "Realtime transcription model",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RealtimeTranscriptionModelResponseDto"
              }
            ]
          },
          "timestampedTranscriptionModel": {
            "description": "Timestamped transcription model",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TimestampedTranscriptionModelResponseDto"
              }
            ]
          }
        },
        "required": [
          "userId",
          "createdAt"
        ]
      },
      "UpdateCustomModelSelectionDto": {
        "type": "object",
        "properties": {
          "responseModelId": {
            "type": "number",
            "description": "Response model ID",
            "nullable": true
          },
          "ttsModelId": {
            "type": "number",
            "description": "TTS model ID",
            "nullable": true
          },
          "realtimeModelId": {
            "type": "number",
            "description": "Realtime model ID",
            "nullable": true
          },
          "realtimeTranscriptionModelId": {
            "type": "number",
            "description": "Realtime transcription model ID",
            "nullable": true
          },
          "timestampedTranscriptionModelId": {
            "type": "number",
            "description": "Timestamped transcription model ID",
            "nullable": true
          }
        }
      },
      "ModelsMessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}